notification formatting logic to other file
